pagination,chaching,etc

+++ alert/models.py
@@ -6,7 +6,7 @@ from django.db import models
 # Create your models here.
 
 class Alert(models.Model):
-    alert_value=models.IntegerField(null=True,blank=True)
+    alert_value=models.FloatField(null=True,blank=True)
     status=models.CharField(max_length=10,null=True,blank=True)
     coin=models.CharField(max_length=20,null=True,blank=True)
     user=models.ForeignKey(User,null=True,blank=True,on_delete=models.CASCADE)
